refactor(embeddings): split clip sidecar store
