+++ src/exomem/embeddings.py
@@ -31,6 +31,8 @@ from typing import NamedTuple
 import numpy as np
 
 from . import accel, index_paths, sidecar_store, vecstore
+from .clip_index import CLIP_DIM, ClipIndex
+from .vector_index_common import vec_gate as _vec_gate
 
 log = logging.getLogger(__name__)
 
@@ -57,8 +59,6 @@ MAX_WORDS_PER_CHUNK = 350
 # by visual content. An EMBEDDER (measurement) like bge — not a captioning VLM —
 # so it stays in-bounds for the pure-substrate server. ViT-B/32 → 512-dim.
 CLIP_MODEL_NAME = "clip-ViT-B-32"
-CLIP_DIM = 512
-
 _MODEL = None
 _MODEL_LOCK = threading.Lock()
 _RERANKER = None
@@ -961,56 +961,6 @@ class _EmbCache(NamedTuple):
     matrix: np.ndarray
 
 
-class _ClipCache(NamedTuple):
-    """ClipIndex's in-memory matrix cache — mirrors `_EmbCache`. `frame_ts[i]` is
-    None for an image row, seconds for a video keyframe row; `paths`/`frame_ts`/
-    `matrix` are parallel arrays."""
-
-    epoch: int
-    generation: int
-    instance: int
-    mtime: float
-    paths: list[str]
-    frame_ts: list[float | None]
-    matrix: np.ndarray
-
-
-def _vec_gate(index, conn: sqlite3.Connection) -> bool:
-    """Shared vec0 policy ladder for EmbeddingIndex/ClipIndex on one connection.
-
-    Duck-typed over the index's vec state (`_vec`, `_vec_failed`, `_vec_ready`,
-    `_vec_quant_synced`): backend gate → extension loadable on this connection →
-    tables created + blob↔vec counts synced (memoized per instance; re-run once
-    more when binary quant turns on later, to synthesize the bin table). Any sync
-    failure retires vec for this instance — the numpy scan serves from then on.
-
-    numpy is the default backend, so this gate returns False (and every dual-write
-    site below is skipped) unless `EXOMEM_VEC_BACKEND=sqlite-vec` opts in. While
-    vec0 is off, the shadow tables drift as the blob tables advance without them;
-    that drift is self-healing — the first opt-in call reaches `ensure_synced`,
-    whose count-mismatch check rebuilds the vec rows from the blobs in pure SQL.
-    """
-    if index._vec_failed or vecstore.backend() == "numpy":
-        return False
-    if not index._vec.try_load(conn):
-        return False
-    quant = vecstore.quant_mode() == "binary"
-    if index._vec_ready is None or (quant and not index._vec_quant_synced):
-        try:
-            index._vec.ensure_synced(conn, quant=quant)
-            index._vec_ready = True
-            if quant:
-                index._vec_quant_synced = True
-        except sqlite3.Error as e:
-            log.warning(
-                "vec sync failed for %s (%s); in-memory scan serves this process",
-                index.path, e,
-            )
-            index._vec_failed = True
-            return False
-    return True
-
-
 class EmbeddingIndex:
     """Per-vault sqlite sidecar holding chunk-level vectors.
 
@@ -1535,434 +1485,6 @@ class EmbeddingIndex:
         path = sidecar_path(vault_root)
         return _sidecar_cache_token(path)
 
-
-class ClipIndex:
-    """Per-vault sqlite sidecar of CLIP image vectors — one vector per image.
-
-    Mirrors EmbeddingIndex (generation-cached matrix, cosine search) but keyed by
-    image file with no chunking. Lives in its own `.clip.sqlite` so the bge text index is
-    untouched and the two evolve independently.
-    """
-
-    def __init__(self, vault_root: Path):
-        self.vault_root = vault_root
-        self.path = clip_sidecar_path(vault_root)
-        self._cache: _ClipCache | None = None
-        self._lock = threading.RLock()
-        # vec0 backend state (see _vec_gate) — mirrors EmbeddingIndex.
-        self._vec = vecstore.SqliteVecStore("images", "vector", CLIP_DIM, "vec_images")
-        self._vec_ready: bool | None = None
-        self._vec_quant_synced = False
-        self._vec_failed = False
-
-    def _connect(self) -> sqlite3.Connection:
-        self.path.parent.mkdir(parents=True, exist_ok=True)
-        conn = sqlite3.connect(self.path)
-        _apply_sidecar_pragmas(conn)
-        # Multi-vector schema: one row per image (frame_ts NULL) OR one row per
-        # video keyframe (frame_ts = seconds). Composite PK keys frames within a
-        # file. NOTE: SQLite treats NULL as DISTINCT in a PRIMARY KEY/UNIQUE index,
-        # so two image rows (same file_path, NULL frame_ts) do NOT collide — every
-        # write path below uses delete-then-insert rather than INSERT OR REPLACE.
-        conn.execute(
-            """
-            CREATE TABLE IF NOT EXISTS images (
-                file_path  TEXT NOT NULL,
-                frame_ts   REAL,
-                vector     BLOB NOT NULL,
-                file_mtime REAL NOT NULL,
-                PRIMARY KEY (file_path, frame_ts)
-            )
-            """
-        )
-        self._migrate_add_frame_ts(conn)
-        _ensure_meta_table(conn, "images", self.path.name)
-        return conn
-
-    @staticmethod
-    def _migrate_add_frame_ts(conn: sqlite3.Connection) -> None:
-        """Upgrade a pre-existing single-vector `images` table in place.
-
-        Old schema: `images(file_path PK, vector, file_mtime)` — no `frame_ts`.
-        SQLite can't ALTER a PRIMARY KEY, so rebuild the table preserving rows
-        (existing image vectors are worth keeping): old rows become image rows
-        (`frame_ts` NULL). Idempotent — no-op once `frame_ts` exists. The CREATE
-        above already made the new table when the sidecar is fresh; this only
-        fires when an OLD table is present.
-        """
-        cols = [r[1] for r in conn.execute("PRAGMA table_info(images)").fetchall()]
-        if "frame_ts" in cols:
-            return
-        with conn:
-            conn.execute(
-                """
-                CREATE TABLE images_new (
-                    file_path  TEXT NOT NULL,
-                    frame_ts   REAL,
-                    vector     BLOB NOT NULL,
-                    file_mtime REAL NOT NULL,
-                    PRIMARY KEY (file_path, frame_ts)
-                )
-                """
-            )
-            conn.execute(
-                "INSERT INTO images_new (file_path, frame_ts, vector, file_mtime) "
-                "SELECT file_path, NULL, vector, file_mtime FROM images"
-            )
-            conn.execute("DROP TABLE images")
-            conn.execute("ALTER TABLE images_new RENAME TO images")
-        log.info("ClipIndex: migrated images table to multi-vector schema (frame_ts)")
-
-    def upsert(self, rel_path: str, vector: np.ndarray, mtime: float) -> None:
-        """Store one image vector (frame_ts NULL). Delete-then-insert: NULL is
-        DISTINCT in the PK, so INSERT OR REPLACE would duplicate the row."""
-        conn = self._connect()
-        try:
-            vec_on = _vec_gate(self, conn)
-            with conn:
-                if vec_on:
-                    # The vec delete carries the SAME partial predicate: replace
-                    # only the image (NULL-ts) row, keep the keyframe rows.
-                    self._vec.dual_delete(
-                        conn, "file_path = ? AND frame_ts IS NULL", (rel_path,)
-                    )
-                conn.execute(
-                    "DELETE FROM images WHERE file_path = ? AND frame_ts IS NULL",
-                    (rel_path,),
-                )
-                conn.execute(
-                    "INSERT INTO images (file_path, frame_ts, vector, file_mtime) "
-                    "VALUES (?, NULL, ?, ?)",
-                    (rel_path, vector.astype(np.float32).tobytes(), mtime),
-                )
-                if vec_on:
-                    self._vec.dual_insert(
-                        conn, "file_path = ? AND frame_ts IS NULL", (rel_path,)
-                    )
-                _bump_meta(conn, "generation")
-                own_epoch, own_gen, own_instance = _read_meta_token(conn)
-        finally:
-            conn.close()
-        # Mirror the partial DELETE: replace only the image (NULL-ts) row, keep any
-        # existing video keyframe rows for this file.
-        self._patch_cache(
-            rel_path,
-            [rel_path],
-            [None],
-            np.asarray(vector, dtype=np.float32).reshape(1, -1),
-            own_epoch,
-            own_gen,
-            own_instance,
-            images_only=True,
-        )
-
-    def upsert_frames(
-        self, rel_path: str, frames: list[tuple[float, np.ndarray]], mtime: float
-    ) -> None:
-        """Replace all vectors for a video with N per-keyframe rows in one txn.
-
-        Clears any prior rows for `rel_path` first — including a stale mean-pooled
-        NULL-ts row from the old single-vector path — then inserts one row per
-        `(timestamp_seconds, vector)`. No-op on an empty frame list (caller soft-skips).
-        """
-        if not frames:
-            return
-        conn = self._connect()
-        try:
-            vec_on = _vec_gate(self, conn)
-            with conn:
-                if vec_on:
-                    self._vec.dual_delete(conn, "file_path = ?", (rel_path,))
-                conn.execute("DELETE FROM images WHERE file_path = ?", (rel_path,))
-                conn.executemany(
-                    "INSERT INTO images (file_path, frame_ts, vector, file_mtime) "
-                    "VALUES (?, ?, ?, ?)",
-                    [
-                        (rel_path, float(ts), vec.astype(np.float32).tobytes(), mtime)
-                        for ts, vec in frames
-                    ],
-                )
-                if vec_on:
-                    self._vec.dual_insert(conn, "file_path = ?", (rel_path,))
-                _bump_meta(conn, "generation")
-                own_epoch, own_gen, own_instance = _read_meta_token(conn)
-        finally:
-            conn.close()
-        # Whole-file DELETE + re-insert → block-replace with the frames, ordered by
-        # timestamp to match the load's `ORDER BY file_path, frame_ts`.
-        ordered = sorted(frames, key=lambda f: f[0])
-        self._patch_cache(
-            rel_path,
-            [rel_path] * len(ordered),
-            [float(ts) for ts, _ in ordered],
-            np.stack([np.asarray(vec, dtype=np.float32) for _, vec in ordered], axis=0),
-            own_epoch,
-            own_gen,
-            own_instance,
-        )
-
-    def delete(self, rel_path: str) -> None:
-        conn = self._connect()
-        try:
-            vec_on = _vec_gate(self, conn)
-            with conn:
-                if vec_on:
-                    self._vec.dual_delete(conn, "file_path = ?", (rel_path,))
-                conn.execute("DELETE FROM images WHERE file_path = ?", (rel_path,))
-                _bump_meta(conn, "generation")
-                own_epoch, own_gen, own_instance = _read_meta_token(conn)
-        finally:
-            conn.close()
-        self._patch_cache(rel_path, [], [], None, own_epoch, own_gen, own_instance)
-
-    def _patch_cache(
-        self,
-        rel_path: str,
-        new_paths: list[str],
-        new_frame_ts: list[float | None],
-        new_vecs: np.ndarray | None,
-        own_epoch: int,
-        own_gen: int,
-        own_instance: int,
-        *,
-        images_only: bool = False,
-    ) -> None:
-        """Splice one file's CLIP rows into the cached matrix (copy-on-write) —
-        ONLY when contiguous with the CURRENT cache: `own_epoch == cached.epoch
-        AND own_instance == cached.instance AND own_gen == cached.generation + 1`.
-        On ANY mismatch, skip the splice ENTIRELY (no content change, no label
-        advance) — see `EmbeddingIndex._patch_cache` for the full corruption
-        rationale (a delayed/out-of-order patch must never overwrite newer rows
-        with stale ones just because its own label looks superficially valid).
-
-        `images_only=True` keeps existing video keyframe rows (frame_ts NOT NULL)
-        and replaces only the image (NULL-ts) row, mirroring `upsert()`'s partial
-        DELETE. Otherwise the file's whole block is replaced. Best-effort: any
-        inconsistency (post-gate) drops the cache to None for a safe full reload
-        next read.
-        """
-        with self._lock:
-            c = self._cache
-            if c is None:
-                return
-            if own_epoch != c.epoch or own_instance != c.instance or own_gen != c.generation + 1:
-                return  # not contiguous with what THIS cache holds -> never splice
-            try:
-                paths, frame_ts, matrix = c.paths, c.frame_ts, c.matrix
-                lo, hi = _file_block(paths, rel_path)
-                if images_only and hi > lo:
-                    keep = [j for j in range(lo, hi) if frame_ts[j] is not None]
-                    block_paths = [paths[j] for j in keep] + list(new_paths)
-                    block_ts = [frame_ts[j] for j in keep] + list(new_frame_ts)
-                    keep_mat = (
-                        matrix[keep] if keep else np.zeros((0, CLIP_DIM), dtype=np.float32)
-                    )
-                    block_parts = [keep_mat]
-                    if new_vecs is not None and new_vecs.shape[0]:
-                        block_parts.append(new_vecs)
-                else:
-                    block_paths = list(new_paths)
-                    block_ts = list(new_frame_ts)
-                    block_parts = (
-                        [new_vecs] if (new_vecs is not None and new_vecs.shape[0]) else []
-                    )
-                out_paths = paths[:lo] + block_paths + paths[hi:]
-                out_ts = frame_ts[:lo] + block_ts + frame_ts[hi:]
-                parts = [matrix[:lo], *block_parts, matrix[hi:]]
-                parts = [p for p in parts if p.shape[0]]
-                out_matrix = (
-                    np.concatenate(parts, axis=0)
-                    if parts
-                    else np.zeros((0, CLIP_DIM), dtype=np.float32)
-                )
-                if not (len(out_paths) == len(out_ts) == out_matrix.shape[0]):
-                    raise ValueError(
-                        f"CLIP splice invariant broken for {rel_path}: "
-                        f"{len(out_paths)} paths / {len(out_ts)} ts / {out_matrix.shape[0]} vecs"
-                    )
-                self._cache = _ClipCache(
-                    c.epoch, own_gen, c.instance, c.mtime, out_paths, out_ts, out_matrix
-                )
-            except Exception as e:  # noqa: BLE001 — self-heal, never break a write
-                log.warning("CLIP matrix splice failed (%s); dropping cache", e)
-                self._cache = None
-
-    def has(self, rel_path: str) -> bool:
-        """True if this file has ANY vector row (image or video). Correct idempotency
-        for IMAGES (one row); for VIDEOS use `has_frames` — a video with only a stale
-        single-vector (frame_ts NULL) row from the old path returns True here but still
-        needs per-keyframe re-indexing."""
-        if not self.path.exists():
-            return False
-        conn = self._connect()
-        try:
-            row = conn.execute(
-                "SELECT 1 FROM images WHERE file_path = ?", (rel_path,)
-            ).fetchone()
-        finally:
-            conn.close()
-        return row is not None
-
-    def has_frames(self, rel_path: str) -> bool:
-        """True if this file has at least one PER-KEYFRAME vector (frame_ts NOT NULL).
-
-        The correct idempotency check for VIDEO: a video carrying only a legacy
-        single-vector row (frame_ts NULL, from the pre-multi-vector path) returns
-        False here, so backfill/worker re-index it per-keyframe instead of skipping it.
-        """
-        if not self.path.exists():
-            return False
-        conn = self._connect()
-        try:
-            row = conn.execute(
-                "SELECT 1 FROM images WHERE file_path = ? AND frame_ts IS NOT NULL",
-                (rel_path,),
-            ).fetchone()
-        finally:
-            conn.close()
-        return row is not None
-
-    def all_vectors(self) -> tuple[list[str], list[float | None], np.ndarray]:
-        """Returns parallel `(file_paths, frame_ts_list, matrix)`, cached until the
-        write generation advances (NOT the mtime — see EmbeddingIndex.all_vectors).
-        frame_ts is None for image rows, a float (seconds) for video keyframe rows."""
-        if not self.path.exists():
-            return [], [], np.zeros((0, CLIP_DIM), dtype=np.float32)
-        c = self._cache  # snapshot once — a concurrent writer may swap/null it
-        served = _try_serve_cached(c, self.path)
-        if served is not None:
-            return served.paths, served.frame_ts, served.matrix
-        with self._lock:
-            # Re-check under the lock (see EmbeddingIndex.all_vectors).
-            c = self._cache
-            served = _try_serve_cached(c, self.path)
-            if served is not None:
-                return served.paths, served.frame_ts, served.matrix
-            loaded = self._load_all_rows()
-            log.info(
-                "CLIP matrix full load: reason=%s rows=%d gen=%d epoch=%d",
-                _reload_reason(c, loaded.epoch, loaded.generation),
-                len(loaded.paths), loaded.generation, loaded.epoch,
-            )
-            self._cache = loaded
-            return loaded.paths, loaded.frame_ts, loaded.matrix
-
-    def unload_cache(self) -> bool:
-        """Drop the resident CLIP matrix cache without deleting sidecar rows."""
-        with self._lock:
-            loaded = self._cache is not None
-            self._cache = None
-            return loaded
-
-    def cache_status(self) -> dict:
-        """Best-effort residency status for this in-memory matrix only."""
-        c = self._cache
-        if c is None:
-            return {"loaded": False, "rows": 0, "bytes": 0}
-        return {
-            "loaded": True,
-            "rows": len(c.paths),
-            "bytes": int(c.matrix.nbytes),
-            "epoch": c.epoch,
-            "generation": c.generation,
-        }
-
-    def _load_all_rows(self) -> _ClipCache:
-        """Full reload → a `_ClipCache`, reading the meta token and rows in ONE
-        `BEGIN` snapshot (see EmbeddingIndex._load_all_rows)."""
-        conn = self._connect()
-        try:
-            conn.execute("BEGIN")
-            try:
-                epoch, gen, instance = _read_meta_token(conn)
-                rows = conn.execute(
-                    "SELECT file_path, frame_ts, vector FROM images "
-                    "ORDER BY file_path, frame_ts"
-                ).fetchall()
-            finally:
-                conn.rollback()  # read-only txn — release the snapshot
-        finally:
-            conn.close()
-        try:
-            mtime = self.path.stat().st_mtime
-        except OSError:
-            mtime = 0.0
-        if not rows:
-            return _ClipCache(
-                epoch, gen, instance, mtime, [], [], np.zeros((0, CLIP_DIM), dtype=np.float32)
-            )
-        paths = [r[0] for r in rows]
-        frame_ts = [r[1] for r in rows]
-        matrix = np.stack([np.frombuffer(r[2], dtype=np.float32) for r in rows], axis=0)
-        return _ClipCache(epoch, gen, instance, mtime, paths, frame_ts, matrix)
-
-    @staticmethod
-    def cache_token(vault_root: Path) -> tuple[int, int, int]:
-        """`(epoch, generation, instance)` for this vault's CLIP sidecar —
-        `(0, 0, 0)` when absent or pre-meta. Read-only (mirrors
-        EmbeddingIndex.cache_token)."""
-        return _sidecar_cache_token(clip_sidecar_path(vault_root))
-
-    def _vec_search(
-        self, query_vec: np.ndarray, k: int
-    ) -> list[tuple[str, float | None, float]] | None:
-        """vec0 KNN, or None when the backend can't serve — mirrors
-        `EmbeddingIndex._vec_search` (no sidecar creation on read, never raises)."""
-        if self._vec_failed or vecstore.backend() == "numpy" or vecstore.load_failed():
-            return None
-        if not self.path.exists():
-            return None
-        try:
-            conn = self._connect()
-            try:
-                if not _vec_gate(self, conn):
-                    return None
-                quant = vecstore.quant_mode() == "binary"
-                pairs = self._vec.knn(conn, query_vec, k, quant=quant)
-                if not pairs:
-                    return []
-                ids = [rid for rid, _ in pairs]
-                placeholders = ",".join("?" * len(ids))
-                rows = conn.execute(
-                    "SELECT rowid, file_path, frame_ts FROM images "
-                    f"WHERE rowid IN ({placeholders})",
-                    ids,
-                ).fetchall()
-                by_id = {r[0]: r for r in rows}
-                return [
-                    (by_id[rid][1], by_id[rid][2], score)
-                    for rid, score in pairs
-                    if rid in by_id
-                ]
-            finally:
-                conn.close()
-        except Exception as e:  # noqa: BLE001 — vec failure must never break search
-            log.warning(
-                "vec search failed for %s (%s); falling back to the in-memory scan",
-                self.path, e,
-            )
-            self._vec_failed = True
-            return None
-
-    def search(self, query_vec: np.ndarray, k: int) -> list[tuple[str, float | None, float]]:
-        """Top-k visual hits: `(file_path, frame_ts, score)` by cosine similarity,
-        sorted by score desc. `frame_ts` is None for images, seconds for video frames.
-        Returns one row per stored vector — a multi-keyframe video yields several rows;
-        callers dedup to best-per-file as needed. Served by the vec0 backend when
-        available (same ladder as `EmbeddingIndex.search`)."""
-        vec_hits = self._vec_search(query_vec, k)
-        if vec_hits is not None:
-            return vec_hits
-        paths, frame_ts, matrix = self.all_vectors()
-        if not paths:
-            return []
-        scores = matrix @ query_vec.astype(np.float32, copy=False)
-        k_eff = min(k, len(scores))
-        if k_eff <= 0:
-            return []
-        top_idx = np.argpartition(-scores, k_eff - 1)[:k_eff]
-        top_idx = top_idx[np.argsort(-scores[top_idx])]
-        return [(paths[i], frame_ts[i], float(scores[i])) for i in top_idx]
 
 
 def vector_backend_active(vault_root: Path) -> bool:
